Removes magithub due to breaking authentication issues.
Recent releases of magithub have switched the authentication
mechanism to not use the native hub tool and instead use a elisp
version. This does not work well and causes magit to break
completely when a strong internet connection is not available.
This is unacceptable and the added functionality from the plugin
is minimal for my use case.