Huge commit -- Full Load-Sequence Refactor
This commit has been in the works for a while, but I finally got
around to actually implementing it. Now all loading of files is
done using a custom macro, which will be improved. To make this
work optimally, the rest of the config was also cleaned up
significantly and moved to separate files. In the future this will
be changed to also load things more lazily.